+++ models/Note.js
@@ -1,16 +1,14 @@
 const mongoose = require('../db/connection')
 
-const noteSchema = new mongoose.Schema({})
-
-
-
-
-const noteSchema = new Schema ({
+const NoteSchema = new mongoose.Schema({
     title: String,
     body: String,
-    author: {UserSchema}
-
+    author: {type: mongoose.Schema.ObjectId, ref: 'Note'}
 })
 
 
-module.exports = mongoose.model('Note', noteSchema)
+
+
+
+
+module.exports = mongoose.model('Note', NoteSchema)
