+++ models/User.js
@@ -1,11 +1,15 @@
 const mongoose = require('../db/connection')
 
-const userSchema = new mongoose.Schema({})
 
-module.exports = mongoose.model('User', userSchema)
 
-const UserSchema = new Schema ({
+const UserSchema = new mongoose.Schema({
     username: String,
     email: String,
-    notes: [NoteSchema]
+    notes: [{  type: mongoose.Schema.ObjectId, ref: 'Note'}]
+    
 })
+
+module.exports = mongoose.model('User', UserSchema)
+
+
+
